[build] move headless rendering files to platform/default
