+++ platform/default/headless_backend_osmesa.cpp
@@ -1,4 +1,4 @@
-#include <mbgl/platform/default/headless_backend.hpp>
+#include <mbgl/gl/headless_backend.hpp>
 #include <mbgl/platform/log.hpp>
 
 #include <GL/osmesa.h>
