+++ platform/default/mbgl/gl/offscreen_view.cpp
@@ -1,4 +1,4 @@
-#include <mbgl/platform/default/offscreen_view.hpp>
+#include <mbgl/gl/offscreen_view.hpp>
 #include <mbgl/gl/context.hpp>
 
 #include <cstring>
