+++ platform/linux/config.cmake
@@ -18,7 +18,7 @@ macro(mbgl_platform_core)
     if(WITH_OSMESA)
         target_sources(mbgl-core
             PRIVATE platform/default/headless_backend_osmesa.cpp
-            PRIVATE platform/default/headless_display.cpp
+            PRIVATE platform/default/mbgl/gl/headless_display.cpp
         )
         target_add_mason_package(mbgl-core PUBLIC mesa)
     elseif(WITH_EGL)
@@ -78,8 +78,11 @@ macro(mbgl_platform_core)
         PRIVATE platform/default/webp_reader.cpp
 
         # Headless view
-        PRIVATE platform/default/headless_backend.cpp
-        PRIVATE platform/default/offscreen_view.cpp
+        PRIVATE platform/default/mbgl/gl/headless_backend.cpp
+        PRIVATE platform/default/mbgl/gl/headless_backend.hpp
+        PRIVATE platform/default/mbgl/gl/headless_display.hpp
+        PRIVATE platform/default/mbgl/gl/offscreen_view.cpp
+        PRIVATE platform/default/mbgl/gl/offscreen_view.hpp
 
         # Thread pool
         PRIVATE platform/default/mbgl/util/default_thread_pool.cpp
