+++ platform/linux/src/headless_backend_glx.cpp
@@ -1,5 +1,5 @@
-#include <mbgl/platform/default/headless_backend.hpp>
-#include <mbgl/platform/default/headless_display.hpp>
+#include <mbgl/gl/headless_backend.hpp>
+#include <mbgl/gl/headless_display.hpp>
 
 #include <mbgl/platform/log.hpp>
