+++ platform/qt/test/headless_backend_qt.cpp
@@ -1,4 +1,4 @@
-#include <mbgl/platform/default/headless_backend.hpp>
+#include <mbgl/gl/headless_backend.hpp>
 
 #include <QApplication>
 #include <QGLContext>
