+++ test/util/offscreen_texture.test.cpp
@@ -2,8 +2,8 @@
 
 #include <mbgl/gl/gl.hpp>
 #include <mbgl/gl/context.hpp>
-#include <mbgl/platform/default/headless_backend.hpp>
-#include <mbgl/platform/default/offscreen_view.hpp>
+#include <mbgl/gl/headless_backend.hpp>
+#include <mbgl/gl/offscreen_view.hpp>
 
 #include <mbgl/util/offscreen_texture.hpp>
